Use a correct URL to the workspace and add a note about signup.
It was not obviousl to me where I had to signup as I clicked directly on
the room link.  Also given I alread has a slack account for a different
workspace trying to add the new workspace was not obvious because
slack.k8s.io is not really a slack workspace, so changed the second URL
to be the workspace URL and not a vanity one.