remove aliases from Chinese pages
The aliases where bogus as we had not been generating them for GH Pages
for a long time.  On top of that when we start generating them they
overwrite the English pages as they missed the /zh/ prefix.

Rather than add /zh/ prefix we just remove them as the links would have
been dead for a long time anyway and this is an opertunity to cleanup