improve example for git token 401
The example used an invalid casing for the token (`-n GitHub`) which doesn't seem to work in real life.

To avoid confusion with the name `admin`, which can make people guess it is a pre-configured value, I've replaced it with `<yourUserName>`, which is what it should be. This implies - correctly in my assumption - that you have to change this value to your git user name.